requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
asyncpg>=0.29.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
alembic>=1.12.1
python-dotenv>=1.0.0
slowapi>=0.1.9
bcrypt>=4.0.1
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0